+++ src/Fl_Message.cxx
@@ -555,7 +555,7 @@ int Fl_Message_Box::handle(int e) {
   switch (e) {
     case FL_KEYBOARD:
     case FL_SHORTCUT:
-      if (Fl::event_key() == 'c' && mods == FL_COMMAND) {
+      if (Fl::event_key() == 'c' && (int)mods == FL_COMMAND) {
         Fl::copy(label(), int(strlen(label())), 1);
         return 1;
       }
